Bug 1318086 - Allow subclasses to override sanitizeArgs, entityFromContext, valueFromContext
